Added resolveRelation function to CHtml and extended resolveValues to support relations
CHtml::resolveRelation evaluates the relation (cascade) for the given model and returns an array with
the related model and an attribute. The relation name can be given in a dot syntax.
For example, if the relation is "author.firstName", this method will return
array('model' => $model->author, 'attribute' => "firstName").
If the last part of a relation cascade is not a relation it is used as the attribute key, otherwise the returned attribtue will be 'primaryKey'.
If the relation cascade is broken in the middle (e.g. $model->comments->author where comments is a MANY_MANY or HAS_MANY relation) it returns
array('model' => null, 'attribute' => null).