Two functions to add button functionality for ajax submissions methods
CHtml class has a method to create a INPUT called "button"
Also it has a method to create a BUTTON called "htmlButton"

On the other hand, there's a method called "ajaxSubmitButton" to create
buttons with ajax calls. This method refers to "ajaxButton" and this
one refers to "button".

My request is to have this feature also in (what you call) htmlButton's.

So there could be a method called "ajaxSubmitHtmlButton" that refers to
"ajaxHtmlButton" and this one to "htmlButton".