issue #531 added asArray property to CActiveFinder
added asArray property to activeFinder to support returning query
results as arrays instead of records.
To fully support this, the API of CActiveRecord needs to be adjusted to
make use of this feature, also some places marked with @todo are not implemented yet.

This commit is ment to be a request for comments.