added test class to test asArray feature
test covers all find methods of CActiveRecord with and without the with
option. So find with CActiveFinder + CActiveRecord is covered nearly
100%.
Currently STAT relations and Composite PK are not explicitly covered.

These tests compare results generated by normal call of CActiveRecord
find*()-methods with the result of find*()-methods in combination
with asArray()
It is not checked if the result is the correct one expected related
to the query criteria.