Update framework/YiiBase.php
In YiiBase::createComponent(), $config array, can take additional element - instanceOf. When this element is specified, createComponent() checks, that new component is instance of specified class/interface (and throws Exception in case of error).  It can be very usefull in some cases - for example when attaching Behavior, and we want be sure, that it implements some interface.

P.S. sory for my bad english.