Fix bug of CRedisCache.php.
Bug of using fread php function. Text from doc:

if the stream is read buffered and it does not represent a plain file, at most one read of up to a number of bytes equal to the chunk size (usually 8192) is made; depending on the previously buffered data, the size of the returned data may be larger than the chunk size.

The result is some cases from redis is read only 8192 byte not all data.