A new property added to the CGridView such that you can set the default column class that will be used by the gridview.
The CGridView now default uses the CDataColumn for any columns that do not have their class defined. This might not always be the wanted functionality. Therefor the new property $defaultColumnClass has been added that allows you to overwrite the default with your own when required. This will then be used for any column that does not have a class defined.